website/requirements.txt
asgiref==3.3.1
Django==3.1.6
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
